week5/requirements.txt
flask==3.0.3
gunicorn==22.0.0
tensorflow-cpu==2.16.1
pillow==10.4.0
numpy==1.26.4